+++ src/components/dialog_config_endpoint.py
@@ -29,22 +29,25 @@ def dialog_config_endpoint(endpoint: Endpoint = None, index: int = None) -> None
     # Values, and default
     name = endpoint.name if endpoint else ""
     url = endpoint.url if endpoint else ""
-    technology_index = technologies_str.index(technology) if endpoint else 0
+    technology_index = technologies_str.index(technology) if endpoint else 3
     base_uri = endpoint.base_uri if endpoint else "http://www.example.org/"
-    ontology_uri = endpoint.ontology_uri if endpoint else "base:shacl"
-    framework_index = frameworks_str.index(ontology_framework) if endpoint else 0
+    ontology_uri = endpoint.ontology_uri if endpoint else "base:ontology"
+    framework_index = frameworks_str.index(ontology_framework) if endpoint else 1
+    metadata_uri = endpoint.metadata_uri if endpoint else "base:metadata"
     username = endpoint.username if endpoint else ""
     password = endpoint.password if endpoint else ""
 
     # Formular
-    endpoint_name = st.text_input('Endpoint name ❗️', value=name, help="Name your endpoint, this is just for you, to recognize it in the list.")
-    endpoint_url = st.text_input('Endpoint URL ❗️', value=url, help="If the endpoint is local, it is propably something like 'http://localhost:9999/', otherwise, see with your endpoint provider.")
-    endpoint_base_uri = st.text_input('Endpoint base URI ❗️', value=base_uri, help="Base URI that will be given to new nodes in the endpoint (plus an ID).")
-    endpoint_technology = st.selectbox('Technology', options=technologies_str, index=technology_index, help="Is your endpoint a Fuseki server? Allegrograph server?")
-    endpoint_ontological_framework = st.selectbox('Select the ontological framework', options=frameworks_str, index=framework_index, help="Only those in the list are supported as of now.")
-    endpoint_ontology_uri = st.text_input('Select the graph in which the ontologycal model lies', value=ontology_uri, help="URI (or shortcut) of the graph containing the ontologycal model; e.g. base:shac.")
-    endpoint_username = st.text_input('Username', value=username, help="In case their is authentication on the endpoint. Leave it empty if not.")
-    endpoint_password = st.text_input('Password', value=password, type='password', help="In case their is authentication on the endpoint. Leave it empty if not.")
+    endpoint_name = st.text_input('Name ❗️', value=name, help="Name to your endpoint, so that you recognize it in the list.")
+    endpoint_url = st.text_input('URL ❗️', value=url, help="If the endpoint is local, it is propably something like 'http://localhost:9999/', see with your endpoint provider otherwise.")
+    endpoint_technology = st.selectbox('Technology ❗️', options=technologies_str, index=technology_index, help="Is your endpoint a technology from the list? Set to \"None\" if it is another one")
+    endpoint_base_uri = st.text_input('Base URI ❗️', value=base_uri, help="Root URI given to new nodes (plus a suffix: an ID), do not forget the trailing \"/\" or \"#\"")
+    endpoint_ontology_uri = st.text_input('Ontology graph URI ❗️', value=ontology_uri, help="URI (or shortcut) of the graph containing the ontologycal model.")
+    endpoint_ontological_framework = st.selectbox('Ontological framework ❗️', options=frameworks_str, index=framework_index, help="As of now, only those in the list are supported.")
+    endpoint_metadata_uri = st.text_input('Metadata graph URI ❗️', value=metadata_uri, help="URI (or shortcut) of the graph containing endpoint metadata.")
+    endpoint_username = st.text_input('Username', value=username, help="In case there is authentication on the endpoint. Leave it empty if not.")
+    endpoint_password = st.text_input('Password', value=password, type='password', help="In case there is authentication on the endpoint. Leave it empty if not.")
+
 
     st.text("")
 
@@ -52,7 +55,7 @@ def dialog_config_endpoint(endpoint: Endpoint = None, index: int = None) -> None
     if st.button('Save'):
 
         # Those are the mandatories fields
-        if endpoint_name and endpoint_url and endpoint_base_uri:
+        if endpoint_name and endpoint_url and endpoint_technology and endpoint_base_uri and endpoint_ontology_uri and endpoint_ontological_framework and endpoint_metadata_uri:
         
             new_endpoint = Endpoint(
                 name=endpoint_name,
@@ -61,6 +64,7 @@ def dialog_config_endpoint(endpoint: Endpoint = None, index: int = None) -> None
                 base_uri=endpoint_base_uri,
                 ontology_uri=endpoint_ontology_uri,
                 ontology_framework=OntologyFramework(endpoint_ontological_framework),
+                metadata_uri=endpoint_metadata_uri,
                 username=endpoint_username,
                 password=endpoint_password
             )