+++ src/components/dialog_config_graph.py
@@ -8,6 +8,9 @@ import lib.state as state
 def dialog_config_graph():
     """Dialog function to provide a formular for the graph creation."""
 
+    # From state
+    endpoint = state.get_endpoint()
+
     # Formular
     graph_name = st.text_input('Name ❗️', value="", help="Will be used as the rdfs:label.")
     graph_uri = st.text_input('Graph URI ❗️', value="base:", help="Value of the graph URI. You can use shortcuts.")
@@ -21,16 +24,12 @@ def dialog_config_graph():
         if graph_name and graph_uri and graph_comment:
 
             # Create triples in default graph
+            triple_class = Triple(graph_uri, 'rdf:type', 'http://www.example.org/Graph')
             triple_name = Triple(graph_uri, 'rdfs:label', f"'{graph_name}'")
             triple_comment = Triple(graph_uri, 'rdfs:comment', f'"""{graph_comment}"""')       
 
-            # Also, in order to make the graph visible in queries, it needs to have at least one triple
-            # So here one dummy triple is created. This solution is in test, let see in the future if this triple is disturbing or not
-            dummy_triple = Triple('_:dummy1', 'base:dummyPredicate', '_:dummy2')
-
             # Insert triples
-            insert([triple_name, triple_comment]) # graph label and comment can't be added to the graph itself by convention
-            insert([dummy_triple], graph=graph_uri, delete_before=False)
+            insert([triple_class, triple_name, triple_comment], graph=endpoint.metadata_uri) # graph label and comment can't be added to the graph itself by convention
 
             # And reset the graphs that are in session, so that, they are fetched on rerun
             state.clear_graphs()